+++ src/basicfmt/basictype_magical.cpp
@@ -99,7 +99,7 @@ void DiskBasicTypeMAGICAL::AdditionalProcessOnMadeDirectory(DiskBasicDirItem *it
 
 	// セクタ名を設定
 	wxUint8 name[32];
-	size_t nlen = sizeof(name);
+	int nlen = (int)sizeof(name);
 	memset(name, 0, nlen);
 	item->GetFileName(name, nlen);
 	sector->Copy(name, nlen);
@@ -137,7 +137,7 @@ bool DiskBasicTypeMAGICAL::AdditionalProcessOnFormatted(const DiskBasicIdentifie
 	if (sector) {
 		sector->Fill(basic->InvertUint8(basic->GetFillCodeOnFAT()));	// invert
 		wxCharBuffer ipl = basic->GetVariousStringParam(wxT("IPLString")).To8BitData();
-		size_t len = ipl.length();
+		int len = (int)ipl.length();
 		if (len > 0) {
 			if (len > 32) len = 32;
 			basic->InvertMem(ipl.data(), len);
